# docs/swagger.yaml
definitions:
  domain.HealthStatusEnum:
    enum:
    - ok
    - error
    type: string
    x-enum-varnames:
    - HealthStatusValueOk
    - HealthStatusValueError
  listopts.PaginationResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  presentation.CreateNoteRequest:
    properties:
      content:
        example: My important note content
        type: string
    type: object
  presentation.ErrorResponse:
    properties:
      message:
        example: user already exists
        type: string
    type: object
  presentation.ListNotesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/presentation.NoteResponse'
        type: array
      pagination:
        $ref: '#/definitions/listopts.PaginationResponse'
    type: object
  presentation.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: ValidPass123!
        type: string
    required:
    - email
    - password
    type: object
  presentation.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  presentation.NoteResponse:
    properties:
      content:
        example: My important note content
        type: string
      created_at:
        example: "2026-02-08T10:30:00Z"
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        example: "2026-02-08T10:30:00Z"
        type: string
      user_id:
        example: 507f1f77bcf86cd799439011
        type: string
    type: object
  presentation.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: ValidPass123!
        type: string
    required:
    - email
    - password
    type: object
  presentation.RegisterResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
    type: object
  presentation.UpdateNoteRequest:
    properties:
      content:
        example: My updated note content
        type: string
    type: object
  response.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
  response.HealthResponse:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.HealthStatusEnum'
        example: ok
    type: object
info:
  contact:
    email: support@momento.com
    name: API Support
  description: API documentation for Momento application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Momento API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password credentials
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/presentation.LoginResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
      summary: Login with email and password
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/presentation.RegisterResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/presentation.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/healthcheck:
    get:
      consumes:
      - application/json
      description: Returns the health status of the application and database connection
      produces:
      - application/json
      responses:
        "200":
          description: Health status response
          schema:
            $ref: '#/definitions/response.HealthResponse'
        "503":
          description: Health status response
          schema:
            $ref: '#/definitions/response.HealthResponse'
      summary: Health check endpoint
      tags:
      - health
  /api/notes:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of notes for the authenticated user
        with sorting options
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - default: created_at
        description: 'Sort field: created_at, updated_at (default: created_at)'
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 'Sort order: asc, desc (default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ListNotesResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: List user notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Creates a new note associated with the authenticated user
      parameters:
      - description: Note content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.NoteResponse'
        "400":
          description: Invalid content
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new note
      tags:
      - notes
  /api/notes/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a note belonging to the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: Retrieves a specific note for the authenticated user by note ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.NoteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a note by ID
      tags:
      - notes
    patch:
      consumes:
      - application/json
      description: Updates the content of a note for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated note content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.NoteResponse'
        "400":
          description: Invalid content
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a note
      tags:
      - notes
  /api/notes/{id}/archive:
    patch:
      consumes:
      - application/json
      description: Archives a note belonging to the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.NoteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Archive a note
      tags:
      - notes
  /api/notes/{id}/restore:
    patch:
      consumes:
      - application/json
      description: Restores a note belonging to the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.NoteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Note not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a note
      tags:
      - notes
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"